requirements.txt
Flask==1.1.2
google-api-python-client==1.12.8
google-cloud-datastore==2.1.0
grpcio==1.29.0
python-dateutil==2.8.1
plotly==4.14.3
numpy==1.19.4
pandas==1.0.3
motuclient==1.8.7
netCDF4==1.5.5
Flask-WTF==0.14.3
WTForms==2.3.3
scikit-learn==0.23.2
joblib==0.16.0
